+++ src/main.c
@@ -6,7 +6,6 @@
 #include "../mods/number/lib/num/struct.h"
 
 #include "../lib/big/header.h"
-#include "../lib/pi/header.h"
 #include "../lib/split/header.h"
 #include "../lib/linear/header.h"
 #include "../lib/union/header.h"
@@ -21,7 +20,7 @@ void time_1()
         TIME_SETUP
         fxd_num_t flt = pi_v1(i);
         TIME_END(t3)
-        printf("\t%.2f", t3/1e9);
+        printf("\t%.2f", (double)t3 / 1e9);
         fxd_num_free(flt);
     }
 }
